+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 Features: 
   - Logo: New `MaisonDeCrowdfundingLogoOnDark` logo.
   - Logo: New `KkbbFlashIllustration` logo.
+  - `DeskLayout`: Add layout component.
+  - `DeskMenu`: Add organism component.
 - Fix: `HeaderMenu`: Fix external item paddings.
 - Fix: `HeaderNav`: Fix border bottom on active nav link.
 - Fix: `TextInput`: Fix `textarea` gradient styles.